+++ CHANGES.md
@@ -1,6 +1,6 @@
 # v0.2 -- UNRELEASED
 
-Added the `binding` step function factory.
+Added the `binding` and `redef` step function factories.
 
 # v0.1 -- 20 Jan 2022