+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 Added the `binding` and `redef` step function factories.
 
+Change `capture-logging` to simply invoke `clojure.tools.logging.test/with-log`; removed `logged-events`.
+
 # v0.1 -- 20 Jan 2022
 
 Initial release.