+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 Added the `binding` and `redef` step function factories.
 
-Change `capture-logging` to simply invoke `clojure.tools.logging.test/with-log`; removed `logged-events`.
+Change `capture-logging` to simply invoke `clojure.tools.logging.test/with-log`; removed `log-events`.
 
 # v0.1 -- 20 Jan 2022